+++ src/core/game_engine.h
@@ -108,7 +108,7 @@ public:
   int target_height() const;
   int width() const;
   int height() const;
-  double view_scale() const;
+  float view_scale() const;
   int target_fps() const;
   const Duration& target_dpf() const;
   const Duration& dpf() const;
@@ -121,7 +121,7 @@ private:
   int target_height_ = 1;
   int width_ = 1;
   int height_ = 1;
-  double view_scale_ = 1.0;
+  float view_scale_ = 1.0f;
   int target_fps_ = 0;
   Duration target_dpf_{};
   Duration dpf_{};
